pyproject.toml
[project]
name = "convert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "protobuf>=5.29.3",
    "pyyaml>=6.0.2",
]
